backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.11.0
pydantic-settings>=2.9.0
python-dotenv>=1.1.0
GitPython>=3.1.0
jira>=3.8.0
fastmcp>=2.6.1
typer>=0.15.0
rich>=13.0.0
httpx>=0.28.0
sqlalchemy>=2.0.0
cachetools>=5.3.0
